OneDrive/Desktop/unimodel/Shree-sakthi/uniagent_backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
mysqlclient>=2.2
google-genai>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.20
pypdf2>=3.0
python-dotenv>=1.0
gunicorn>=21.2
